+++ AgreementMaker-OSGi/AgreementMaker-API/src/main/java/am/api/config/IsConfigurable.java
@@ -1,0 +1,10 @@
+package am.api.config;
+
+import java.util.List;
+import java.util.Map;
+import java.util.Properties;
+
+public interface IsConfigurable {
+    List<Enum> getConfigurationKeys();
+    void configure(Map<Enum, String> properties);
+}
