+++ AgreementMaker-OSGi/AgreementMaker-Core/pom.xml
@@ -174,6 +174,12 @@
 			<version>0.0.1-SNAPSHOT</version>
 			<scope>provided</scope>
 		</dependency>
+		<dependency>
+			<groupId>org.inferred</groupId>
+			<artifactId>freebuilder</artifactId>
+			<version>1.10.5</version>
+			<optional>true</optional>
+		</dependency>
 
 		<dependency>
 			<groupId>edu.uic.cs.advis.am</groupId>
@@ -266,18 +272,6 @@
 			<groupId>org.apache.commons</groupId>
 			<artifactId>commons-compress</artifactId>
 			<version>1.4.1</version>
-		</dependency>
-
-		<dependency>
-			<groupId>secondstring</groupId>
-			<artifactId>secondstring</artifactId>
-			<version>20120620</version>
-		</dependency>
-
-		<dependency>
-			<groupId>simpack</groupId>
-			<artifactId>simpack</artifactId>
-			<version>0.91</version>
 		</dependency>
 
 		<dependency>
