+++ AgreementMaker-OSGi/AgreementMaker-StringUtil/src/main/java/am/stringutil/Levenstein.java
@@ -1,0 +1,7 @@
+package am.stringutil;
+
+/**
+ * Created by marin on 10/21/16.
+ */
+public class Levenstein {
+}
