+++ AgreementMaker-OSGi/AgreementMaker-UI/src/main/java/am/visualization/WordNetLookup/WordNetLookupPanel.java
@@ -39,22 +39,18 @@ import java.io.IOException;
 import javax.imageio.ImageIO;
 import javax.swing.BorderFactory;
 import javax.swing.GroupLayout;
-import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JFrame;
 import javax.swing.JLabel;
 import javax.swing.JOptionPane;
-import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JTextField;
 import javax.swing.JTextPane;
 import javax.swing.SwingUtilities;
 import javax.swing.text.html.HTMLDocument;
 
-import am.Utility;
 import am.app.mappingEngine.StringUtil.PorterStemmer;
 import am.ui.UIUtility;
-import am.ui.api.AMTab;
 import am.ui.api.impl.AMTabSupportPanel;
 import am.visualization.graphviz.GraphViz;
 import am.visualization.graphviz.wordnet.NavigableImagePanel;
